move deprecated content to _deprecated
